feat: Revamp smart prompting and instruction server architecture with modular sub-servers
