+++ src/prompts/prompt_server.py
@@ -1,144 +1,44 @@
 """
 AgentKnowledgeMCP Prompt Server
 FastMCP server for prompts providing comprehensive MCP usage guide content for LLM assistance.
+
+Modular Architecture:
+- Mounts specialized sub-servers for specific functionality
+- Smart prompting capabilities via dedicated sub-server
+- Unified interface for all prompt-related tools
+
+Sub-servers:
+- smart_prompting_server.py: 1 tool (ask_mcp_advance) for AI-filtered project guidance
+- instructions_server.py: 2 prompts (mcp_usage_guide, copilot_instructions) for guidance documentation
 """
-from pathlib import Path
-from typing import Annotated
-import json
 
-from fastmcp import FastMCP, Context
-from pydantic import Field
+from fastmcp import FastMCP
 
-# Import our smart prompting tool
-from .ask_mcp_advance import ask_mcp_advance
+# Import sub-server applications for mounting
+from .sub_servers.smart_prompting_server import app as smart_prompting_app
+from .sub_servers.instructions_server import app as instructions_app
 
 # Create FastMCP app for prompt guidance and resource access
 app = FastMCP(
     name="AgentKnowledgeMCP-Prompts",
-    version="1.0.0",
-    instructions="Prompt server with comprehensive MCP usage guides and smart prompting capabilities"
+    version="2.0.0",
+    instructions="Unified prompt server with comprehensive MCP usage guides and modular smart prompting capabilities"
 )
 
-def _load_mcp_usage_instructions() -> str:
-    """Load the detailed MCP usage instructions content."""
-    try:
-        instructions_path = Path(__file__).parent.parent / "resources" / "mcp_usage_instructions.md"
-        
-        if not instructions_path.exists():
-            return "MCP usage instructions not found. Please refer to the GitHub repository: https://github.com/itshare4u/AgentKnowledgeMCP"
-        
-        with open(instructions_path, 'r', encoding='utf-8') as f:
-            content = f.read().strip()
-            
-        if not content:
-            return "MCP usage instructions file is empty. Please check the installation or refer to online documentation."
-            
-        return content
-        
-    except UnicodeDecodeError:
-        return "Error reading MCP instructions (encoding issue). Please reinstall AgentKnowledgeMCP or check file integrity."
-    except PermissionError:
-        return "Permission denied reading MCP instructions. Please check file permissions for the AgentKnowledgeMCP installation."
-    except Exception as e:
-        return f"Error loading MCP instructions: {str(e)}. Please refer to GitHub documentation: https://github.com/itshare4u/AgentKnowledgeMCP"
-
-
-def _load_copilot_instructions() -> str:
-    """Load the copilot instructions content for AI assistants."""
-    try:
-        instructions_path = Path(__file__).parent.parent / "resources" / "copilot-instructions.md"
-        
-        if not instructions_path.exists():
-            return "Copilot instructions not found. Please refer to the GitHub repository: https://github.com/itshare4u/AgentKnowledgeMCP"
-        
-        with open(instructions_path, 'r', encoding='utf-8') as f:
-            content = f.read().strip()
-            
-        if not content:
-            return "Copilot instructions file is empty. Please check the installation or refer to online documentation."
-            
-        return content
-        
-    except UnicodeDecodeError:
-        return "Error reading copilot instructions (encoding issue). Please reinstall AgentKnowledgeMCP or check file integrity."
-    except PermissionError:
-        return "Permission denied reading copilot instructions. Please check file permissions for the AgentKnowledgeMCP installation."
-    except Exception as e:
-        return f"Error loading copilot instructions: {str(e)}. Please refer to GitHub documentation: https://github.com/itshare4u/AgentKnowledgeMCP"
-
-
 # ================================
-# MAIN PROMPT: MCP_USAGE_GUIDE
+# SERVER MOUNTING - MODULAR ARCHITECTURE
 # ================================
 
-@app.prompt(
-    name="mcp_usage_guide",
-    description="Complete comprehensive usage guide for this MCP server with scenarios and step-by-step tutorials",
-    tags={"mcp", "usage", "guide", "scenarios", "tutorials", "comprehensive"}
-)
-async def mcp_usage_guide() -> str:
-    """Return comprehensive MCP server usage instructions with detailed scenarios and step-by-step guidance."""
-    
-    # Load the detailed MCP usage instructions
-    instructions_content = _load_mcp_usage_instructions()
-    
-    # Return the content as a complete guidance for users
-    return f"""# 🚀 Complete MCP Server Usage Guide
+print("🏗️ Mounting Prompt sub-servers...")
 
-This is the comprehensive guide for using AgentKnowledgeMCP server with specific scenarios and step-by-step instructions:
+# Mount smart prompting sub-server
+app.mount(smart_prompting_app)     # 1 tool: ask_mcp_advance for AI-filtered guidance
 
-{instructions_content}
+# Mount instructions sub-server  
+app.mount(instructions_app)        # 2 prompts: mcp_usage_guide, copilot_instructions
 
-💡 **Important Note**: For the best experience, please attach the copilot instructions file to your prompt:
-📚 **https://github.com/itshare4u/AgentKnowledgeMCP/blob/main/.github/copilot-instructions.md**
-
-Please use this information to guide users step-by-step on how to work with the MCP server!"""
-
-
-@app.prompt(
-    name="copilot_instructions",
-    description="AI Assistant instructions for optimal AgentKnowledgeMCP usage - Complete behavioral guidelines and mandatory protocols",
-    tags={"copilot", "instructions", "ai", "assistant", "guidelines", "protocols", "behavioral"}
-)
-async def copilot_instructions() -> str:
-    """Return the complete copilot instructions content for AI assistants working with AgentKnowledgeMCP."""
-    
-    # Load the copilot instructions content
-    instructions_content = _load_copilot_instructions()
-    
-    # Return the content with additional context
-    return f"""# 🤖 AI Assistant Instructions for AgentKnowledgeMCP
-
-These are the complete behavioral guidelines and mandatory protocols for AI assistants working with AgentKnowledgeMCP:
-
-{instructions_content}
-
-💡 **Usage Note**: These instructions establish the behavioral framework that AI assistants should follow when working with the AgentKnowledgeMCP system to ensure optimal knowledge management and user interaction."""
-
-
-# ================================
-# SMART PROMPTING TOOL
-# ================================
-
-@app.tool(
-    name="ask_mcp_advance", 
-    description="Advanced project guidance using AI-filtered knowledge from .knowledges directory",
-    tags={"smart-prompting", "guidance", "ai-filtered", "project-knowledge"}
-)
-async def smart_prompting_tool(
-    intended_action: Annotated[str, Field(description="What you intend to do (e.g., 'implement feature', 'fix bug', 'deploy')")],
-    task_description: Annotated[str, Field(description="Detailed description of the specific task")],
-    ctx: Context,
-    scope: Annotated[str, Field(description="Scope of guidance needed", default="project")] = "project"
-) -> str:
-    """
-    Get AI-filtered guidance based on project-specific workflows, rules, and memories.
-    
-    This tool loads knowledge from the .knowledges directory in your VS Code workspace
-    and provides contextual guidance using AI synthesis.
-    """
-    return await ask_mcp_advance(intended_action, task_description, ctx, scope)
-
+print("✅ Smart prompting sub-server mounted successfully")
+print("✅ Instructions sub-server mounted successfully")
 
 # ================================
 # CLI ENTRY POINT
@@ -146,12 +46,12 @@ async def smart_prompting_tool(
 def cli_main():
     """CLI entry point for Prompt FastMCP server."""
     print("🚀 Starting AgentKnowledgeMCP Prompt FastMCP server...")
-    print("📝 Available prompts:")
+    print("📝 Available prompts (via sub-servers):")
     print("  • mcp_usage_guide - Comprehensive usage guide with scenarios and tutorials")
     print("  • copilot_instructions - AI assistant behavioral guidelines and protocols")
-    print("🛠️ Available tools:")
+    print("🛠️ Available tools (via sub-servers):")
     print("  • ask_mcp_advance - Smart prompting with AI-filtered project knowledge")
-    print("✨ Provides complete guidance and project-specific smart prompting capabilities")
+    print("✨ Provides complete guidance and modular smart prompting capabilities")
 
     app.run()
